requirements.txt
flask
torch
torchvision
Pillow
numpy
scikit-learn